requirements.txt
alembic==1.11.1
attrs==24.2.0
blinker==1.6.2
certifi==2024.8.30
click==8.1.3
escape==1.1
et_xmlfile==2.0.0
flasgger==0.9.7.1
Flask==2.3.2
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
flask-cors==5.0.0
flask_session
fuzzywuzzy==0.18.0
GeoAlchemy2==0.15.2
geobuf==2.0.0
geopandas==1.0.1
google-auth
greenlet==2.0.2
gunicorn==23.0.0                               
haversine==2.8.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
fiona==1.10.1
Jinja2==3.1.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Levenshtein==0.26.1
Mako==1.2.4
MarkupSafe==2.1.3
mistune==3.0.2
numpy==2.0.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
protobuf==3.20.3
psycopg2==2.9.6
PyJWT==2.9.0
pyogrio==0.10.0
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.26.1
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
rpds-py==0.20.1
requests
shapely==2.0.6
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
tzdata==2024.2
Werkzeug==2.3.5
zipp==3.15.0